Fix a bug during the import process.
The import module would not import courses in which the importer was enroller.
The import module assumed that any row starting with a space in the first
column was a continuation of the class in the previous record. However, if the
importer is enrolled in a course, this first column will contain a space as
well. Thus, the second column, the CRN, should be used, as it is present for a
new record regardless of the enrollment status.

In addition to this fix, additional debugging statements were added for cases
when a line is skipped.