Add active schedule support.
Add support for setting the active schedule. The active schedule is the
schedule to which courses found in the search interface will be added, unless
otherwise specified.

In addition, on the list of schedules page, if a user does not have any
schedules, display a message instructing them to create a schedule, rather
than displaying an empty table.