Set active schedule when creating a schedule.
When a new schedule is created, it is automatically made the active schedule
if no other schedule exists. If another schedule already exists, a checkbox
is displayed, allowing the user to make this new schedule the active schedule.

When a schedule is removed, the first schedule in the list of schedules is set
as the active schedule, rather than whichever is returned first by the database.