Change _schedule_display partial display increment.
Change the _schedule_display partial increment so that each row
represents five minutes instead of ten. However, only display dashed
lines in the schedule on ten minute increments, as before. This fixes
two bugs:
   1) If one course ends and another starts within the same ten minute
      period, then there would be an error in the display of the
      courses.
   2) Courses not starting on a ten minute barrier were not being
      displayed. Now courses must start on a five minute barrier. If
      this is not the case, then the course start time will be rounded
      down to a five minute barrier.

In addition, the end time was already rounded up to an increment of the
barrier time period, but in a different spot. This commit moves that
rounding process to the same location as the fix for #2, so rounding
does not take place in two separate places.