Modify the _schedule_display partial sort order.
Modify the _schedule_display partial so that it sorts courses in the
sidebar based on their department abbreviation, course number, and
section. Assignment of colors is done after this sorting process, so
colors are never repeated next to each other in long lists.

In addition, modify the color assignment process so that it can be
completed after the creation of meeting_times and calendar, as there was
no real reason each meeting_time needed to contain a copy of the color.