Add multiple credit hour counts, if applicable.
Add multiple credit hour counts if a schedule contains more than one
course section belonging to the same course. This allows people to add
multiple sections to a schedule, in order to decide which they like
best, while still getting an accurate count of the credit hours in the
schedule.