Fix module map loading bug.
Previously, the module map would not be loaded before classes were
loaded using the __autoload() function. This would prompt a class not
found error, which would result in the regeneration of the module map,
after which the module map was loaded. This loads the module map first,
so the module map generation only happens if needed.