Associate credit hours with sections, not courses.
All sections in a course do not necessarily have the same number of
credit hours. This is common with labs, where someone may sign up for
the lecture section and the lab section, but only the lecture section
has credit hours. If both had credit hours, then the course would be
double-counted.